web/requirements.txt
# Requirements untuk DEPLOY
Flask==3.1.1
scikit-learn==1.7.0
scikit-image==0.25.2
numpy==2.2.6
joblib==1.5.1
opencv-python==4.11.0.86
Pillow==11.2.1